Changed datatypes for lat and lng attributes of Spot table

+++ backend/db/migrations/1-create-spot.js
@@ -42,11 +42,11 @@ module.exports = {
         allowNull: false
       },
       lat: {
-        type: Sequelize.DECIMAL(5,10),
+        type: Sequelize.FLOAT,
         allowNull: false,
       },
       lng: {
-        type: Sequelize.DECIMAL(5,10),
+        type: Sequelize.FLOAT,
         allowNull: false,
       },
       name: {
